Refactor for jest test config
