+++ src/app.ts
@@ -1,0 +1,29 @@
+import cors from 'cors';
+import helmet from 'helmet';
+import morgan from 'morgan';
+import express from 'express';
+import bodyParser from 'body-parser';
+
+import routes from './routes';
+import config from './config/config';
+import notFoundHandler from './middlewares/notFoundHandler';
+import genericErrorHandler from './middlewares/genericErrorHandler';
+
+const { name, version } = config;
+const app: express.Application = express();
+
+app.locals.name = name;
+app.locals.version = version;
+
+app.use(cors());
+app.use(helmet());
+app.use(morgan('dev'));
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
+
+app.use('/api', routes);
+
+app.use(genericErrorHandler);
+app.use(notFoundHandler);
+
+export default app;
